Comit before Holders Classes
Added core artifacts, enemies, and much other.

+++ Assets/Scripts/WordsPhrase/Phrases/Equipment/Armor/Armor.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 public class Armor : Equipment
 {
     [SerializeField] private int _defendPercent;
-    [SerializeField] private float _healthRegeneration;
-    [SerializeField] private float _manaRegeneration;
+    [SerializeField] private int _healthRegeneration;
+    [SerializeField] private int _manaRegeneration;
 
     public int DefendPercent => _defendPercent;
     public float HealthRegeneration => _healthRegeneration;
