+++ Assets/Scripts/WordsPhrase/Phrases/Equipment/Magic/Magic.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public abstract class Magic : Equipment
 {
-    [SerializeField] private float _manaCost;
+    [SerializeField] private int _manaCost;
     [SerializeField] private float _castDelay;
 
     public float CastDelay => _castDelay;
